Update README.md
This exercise aims to teach the manipulation of numeric data types in a MySQL database. It involves creating a database called **cm_devices** for a fictional company, **CM Mobiles**, specializing in selling mobile devices. The exercise is divided into two main parts: creating a **devices** table to store information about the devices (ID, name, price) and creating a **stock** table to manage inventory quantities and total costs. Through these tasks, the user learns to define columns with appropriate data types (`INT` for integers and `DECIMAL` for decimal values), execute essential SQL commands such as `CREATE DATABASE`, `CREATE TABLE`, `USE`, `SHOW TABLES`, and `SHOW COLUMNS`, and structure a database according to business requirements. This exercise is designed to be performed on the Coursera platform, where it helps reinforce the understanding of SQL fundamentals and the use of numeric data types in a practical context